+++ Frieght.Api/Dtos/LoadDto.cs
@@ -7,6 +7,7 @@ public record LoadDto
 (
     int Id,
     string UserId,
+    // string ShipperUserId,
     ShipperDto CreatedBy,
     string Origin,
     string destination ,
@@ -26,7 +27,7 @@ public record LoadDto
 public record CreateLoadDto
 (
     string UserId,
-    ShipperDto CreatedBy,
+    string ShipperUserId,
     [Required][StringLength(100)] string Origin,
     [Required][StringLength(100)] string destination,
     DateTime PickupDate,
@@ -56,10 +57,24 @@ public record UpdateLoadDto
     string ModifiedBy
 );
 
-public record ShipperDto
-(
-    string? Id,
-    string? Email,  
-    string? FirstName,
-    string? LastName
-);
+
+
+// public class LoadDto {
+//     public int Id { get; set; }
+//     public string UserId { get; set; }
+//     public string ShipperUserId { get; set; }
+//     public dynamic Shipper { get; set; }
+//     public string Origin { get; set; }
+//     public string destination { get; set; }
+//     public DateTime PickupDate { get; set; }
+//     public DateTime DeliveryDate { get; set; }
+//     public string Commodity { get; set; }
+//     public double Weight { get; set; }
+//     public decimal OfferAmount { get; set; }
+//     public string LoadDetails { get; set; }
+//     public string LoadStatus { get; set; }
+//     public int? AcceptedBidId { get; set; }
+//     public DateTime? Created { get; set; }
+//     public DateTime? Modified { get; set; }
+//     public string ModifiedBy { get; set; }
+// }