+++ Frieght.Api/Entities/Load.cs
@@ -1,5 +1,4 @@
 ﻿using System.ComponentModel.DataAnnotations;
-using Frieght.Api.Dtos;
 
 namespace Frieght.Api.Entities;
 
@@ -8,7 +7,8 @@ public class Load
     public int Id { get; set; }
     [Required]
     public required string UserId { get; set; }
-    public ShipperDto CreatedBy { get; set; }
+    public required string ShipperUserId { get; set; }
+    public Shipper Shipper { get; set; }
     [Required]
     [StringLength(100)]
     public required string Origin { get; set; }