+++ Frieght.Api/Infrastructure/Data/Migrations/20240317191256_FixCreatedByInLoad.Designer.cs
@@ -3,6 +3,7 @@ using System;
 using Frieght.Api.Infrastructure;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Infrastructure;
+using Microsoft.EntityFrameworkCore.Migrations;
 using Microsoft.EntityFrameworkCore.Storage.ValueConversion;
 using Npgsql.EntityFrameworkCore.PostgreSQL.Metadata;
 
@@ -11,9 +12,11 @@ using Npgsql.EntityFrameworkCore.PostgreSQL.Metadata;
 namespace Frieght.Api.Infrastructure.Data.Migrations
 {
     [DbContext(typeof(FrieghtDbContext))]
-    partial class FrieghtDbContextModelSnapshot : ModelSnapshot
+    [Migration("20240317191256_FixCreatedByInLoad")]
+    partial class FixCreatedByInLoad
     {
-        protected override void BuildModel(ModelBuilder modelBuilder)
+        /// <inheritdoc />
+        protected override void BuildTargetModel(ModelBuilder modelBuilder)
         {
 #pragma warning disable 612, 618
             modelBuilder
@@ -21,6 +24,31 @@ namespace Frieght.Api.Infrastructure.Data.Migrations
                 .HasAnnotation("Relational:MaxIdentifierLength", 63);
 
             NpgsqlModelBuilderExtensions.UseIdentityByDefaultColumns(modelBuilder);
+
+            modelBuilder.Entity("Frieght.Api.Dtos.ShipperDto", b =>
+                {
+                    b.Property<string>("Id")
+                        .HasColumnType("text");
+
+                    b.Property<string>("CompanyName")
+                        .HasColumnType("text");
+
+                    b.Property<string>("DOTNumber")
+                        .HasColumnType("text");
+
+                    b.Property<string>("Email")
+                        .HasColumnType("text");
+
+                    b.Property<string>("FirstName")
+                        .HasColumnType("text");
+
+                    b.Property<string>("LastName")
+                        .HasColumnType("text");
+
+                    b.HasKey("Id");
+
+                    b.ToTable("ShipperDto");
+                });
 
             modelBuilder.Entity("Frieght.Api.Entities.Bid", b =>
                 {
@@ -121,6 +149,9 @@ namespace Frieght.Api.Infrastructure.Data.Migrations
                     b.Property<DateTime>("Created")
                         .HasColumnType("timestamp with time zone");
 
+                    b.Property<string>("CreatedById")
+                        .HasColumnType("text");
+
                     b.Property<DateTime>("DeliveryDate")
                         .HasColumnType("timestamp with time zone");
 
@@ -152,10 +183,6 @@ namespace Frieght.Api.Infrastructure.Data.Migrations
                     b.Property<DateTime>("PickupDate")
                         .HasColumnType("timestamp with time zone");
 
-                    b.Property<string>("ShipperUserId")
-                        .IsRequired()
-                        .HasColumnType("text");
-
                     b.Property<string>("UserId")
                         .IsRequired()
                         .HasColumnType("text");
@@ -170,50 +197,18 @@ namespace Frieght.Api.Infrastructure.Data.Migrations
 
                     b.HasKey("Id");
 
-                    b.HasIndex("ShipperUserId");
+                    b.HasIndex("CreatedById");
 
                     b.ToTable("Loads");
                 });
 
-            modelBuilder.Entity("Frieght.Api.Entities.Shipper", b =>
-                {
-                    b.Property<string>("UserId")
-                        .HasColumnType("text");
-
-                    b.Property<string>("CompanyName")
-                        .HasColumnType("text");
-
-                    b.Property<string>("DOTNumber")
-                        .HasColumnType("text");
-
-                    b.Property<string>("Email")
-                        .HasColumnType("text");
-
-                    b.Property<string>("FirstName")
-                        .HasColumnType("text");
-
-                    b.Property<string>("LastName")
-                        .HasColumnType("text");
-
-                    b.HasKey("UserId");
-
-                    b.ToTable("Shipper");
-                });
-
             modelBuilder.Entity("Frieght.Api.Entities.Load", b =>
                 {
-                    b.HasOne("Frieght.Api.Entities.Shipper", "Shipper")
-                        .WithMany("Loads")
-                        .HasForeignKey("ShipperUserId")
-                        .OnDelete(DeleteBehavior.Cascade)
-                        .IsRequired();
+                    b.HasOne("Frieght.Api.Dtos.ShipperDto", "CreatedBy")
+                        .WithMany()
+                        .HasForeignKey("CreatedById");
 
-                    b.Navigation("Shipper");
-                });
-
-            modelBuilder.Entity("Frieght.Api.Entities.Shipper", b =>
-                {
-                    b.Navigation("Loads");
+                    b.Navigation("CreatedBy");
                 });
 #pragma warning restore 612, 618
         }