+++ Frieght.Api/Infrastructure/Data/Migrations/20240317202548_FixShipperInfo.Designer.cs
@@ -3,6 +3,7 @@ using System;
 using Frieght.Api.Infrastructure;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Infrastructure;
+using Microsoft.EntityFrameworkCore.Migrations;
 using Microsoft.EntityFrameworkCore.Storage.ValueConversion;
 using Npgsql.EntityFrameworkCore.PostgreSQL.Metadata;
 
@@ -11,9 +12,11 @@ using Npgsql.EntityFrameworkCore.PostgreSQL.Metadata;
 namespace Frieght.Api.Infrastructure.Data.Migrations
 {
     [DbContext(typeof(FrieghtDbContext))]
-    partial class FrieghtDbContextModelSnapshot : ModelSnapshot
+    [Migration("20240317202548_FixShipperInfo")]
+    partial class FixShipperInfo
     {
-        protected override void BuildModel(ModelBuilder modelBuilder)
+        /// <inheritdoc />
+        protected override void BuildTargetModel(ModelBuilder modelBuilder)
         {
 #pragma warning disable 612, 618
             modelBuilder
@@ -121,6 +124,10 @@ namespace Frieght.Api.Infrastructure.Data.Migrations
                     b.Property<DateTime>("Created")
                         .HasColumnType("timestamp with time zone");
 
+                    b.Property<string>("CreatedByUserId")
+                        .IsRequired()
+                        .HasColumnType("text");
+
                     b.Property<DateTime>("DeliveryDate")
                         .HasColumnType("timestamp with time zone");
 
@@ -152,10 +159,6 @@ namespace Frieght.Api.Infrastructure.Data.Migrations
                     b.Property<DateTime>("PickupDate")
                         .HasColumnType("timestamp with time zone");
 
-                    b.Property<string>("ShipperUserId")
-                        .IsRequired()
-                        .HasColumnType("text");
-
                     b.Property<string>("UserId")
                         .IsRequired()
                         .HasColumnType("text");
@@ -170,7 +173,7 @@ namespace Frieght.Api.Infrastructure.Data.Migrations
 
                     b.HasKey("Id");
 
-                    b.HasIndex("ShipperUserId");
+                    b.HasIndex("CreatedByUserId");
 
                     b.ToTable("Loads");
                 });
@@ -202,18 +205,13 @@ namespace Frieght.Api.Infrastructure.Data.Migrations
 
             modelBuilder.Entity("Frieght.Api.Entities.Load", b =>
                 {
-                    b.HasOne("Frieght.Api.Entities.Shipper", "Shipper")
-                        .WithMany("Loads")
-                        .HasForeignKey("ShipperUserId")
+                    b.HasOne("Frieght.Api.Entities.Shipper", "CreatedBy")
+                        .WithMany()
+                        .HasForeignKey("CreatedByUserId")
                         .OnDelete(DeleteBehavior.Cascade)
                         .IsRequired();
 
-                    b.Navigation("Shipper");
-                });
-
-            modelBuilder.Entity("Frieght.Api.Entities.Shipper", b =>
-                {
-                    b.Navigation("Loads");
+                    b.Navigation("CreatedBy");
                 });
 #pragma warning restore 612, 618
         }