+++ Changelog.md
@@ -2,3 +2,4 @@
 
 * [Task 43] - Finish Registration and integrate it with the database.
 * [Task 39] - Finish Auto-deployent setup for frieght.api and make it accessible.
+* [Task 39] -  Add shipper info into load entity